pyproject.toml
[tool.poetry]
name = "image_captions"
version = "0.1.0"
description = ""
authors = ["Pablo Campillo <dev@pablocampillo.pro>"]
readme = "README.md"
packages = [{include = "image_captions", from = "src"}]


[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.35.2"
pillow = "^10.1.0"
tornado = "^6.3.3"
torch = {version = "^2.1.1+cu118", source = "pytorch-gpu-src"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
matplotlib = "^3.8.2"
pytest-tornasync = "^0.6.0.post2"


[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
image_captions = "image_captions.app:app"